生產部署/偉大/python-bot/requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Telegram Bot
python-telegram-bot==20.7
aiohttp==3.9.1

# 數據庫
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Firebase
firebase-admin==6.2.0

# 加密安全
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# 數據處理
pandas==2.1.4
numpy==1.25.2

# HTTP 請求
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# 環境變數
python-dotenv==1.0.0

# 日誌
loguru==0.7.2

# 支付相關
tronpy==0.4.0

# 工具
python-multipart==0.0.6
